peripheral array
