+++ svd-encoder/CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## Unreleased
+
+- Add array support for peripherals
+
 ## [v0.11.0] - 2021-10-02
 - Splitted from `svd-parser`
