+++ svd-encoder/src/cluster.rs
@@ -7,8 +7,8 @@ impl Encode for Cluster {
 
     fn encode(&self) -> Result<Element, EncodeError> {
         match self {
-            Cluster::Single(i) => i.encode(),
-            Cluster::Array(i, a) => {
+            Self::Single(i) => i.encode(),
+            Self::Array(i, a) => {
                 let mut e = Element::new("cluster");
                 e.merge(&a.encode()?);
                 e.merge(&i.encode()?);
