+++ svd-parser/src/cluster.rs
@@ -26,9 +26,9 @@ impl Parse for Cluster {
                 }
             }
 
-            Ok(Cluster::Array(info, array_info))
+            Ok(Self::Array(info, array_info))
         } else {
-            Ok(Cluster::Single(info))
+            Ok(Self::Single(info))
         }
     }
 }
