+++ svd-rs/src/derive_from.rs
@@ -1,7 +1,7 @@
 //! Implementations of DeriveFrom, setting non-explicit fields.
 use crate::{
-    Cluster, ClusterInfo, EnumeratedValues, Field, FieldInfo, Peripheral, Register, RegisterInfo,
-    RegisterProperties,
+    Cluster, ClusterInfo, EnumeratedValues, Field, FieldInfo, Peripheral, PeripheralInfo, Register,
+    RegisterInfo, RegisterProperties,
 };
 
 /// Fill empty fields of structure with values of other structure
@@ -37,7 +37,7 @@ impl DeriveFrom for EnumeratedValues {
     }
 }
 
-impl DeriveFrom for Peripheral {
+impl DeriveFrom for PeripheralInfo {
     fn derive_from(&self, other: &Self) -> Self {
         let mut derived = self.clone();
         derived.group_name = derived.group_name.or_else(|| other.group_name.clone());
@@ -91,6 +91,23 @@ impl DeriveFrom for FieldInfo {
             .modified_write_values
             .or(other.modified_write_values);
         derived
+    }
+}
+
+impl DeriveFrom for Peripheral {
+    fn derive_from(&self, other: &Self) -> Self {
+        match (self, other) {
+            (Self::Single(info), Self::Single(other_info)) => {
+                Self::Single(info.derive_from(other_info))
+            }
+            (Self::Single(info), Self::Array(other_info, other_dim)) => {
+                Self::Array(info.derive_from(other_info), other_dim.clone())
+            }
+            (Self::Array(info, dim), Self::Single(other_info))
+            | (Self::Array(info, dim), Self::Array(other_info, _)) => {
+                Self::Array(info.derive_from(other_info), dim.clone())
+            }
+        }
     }
 }
